# attached_assets/Bot-Host-Hub/Bot-Host-Hub/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface AccessKeyInput {
  /** @minLength 1 */
  key: string;
}

export interface BotUpload {
  file: Blob;
}

export type HostedBotSummaryStatus = typeof HostedBotSummaryStatus[keyof typeof HostedBotSummaryStatus];


export const HostedBotSummaryStatus = {
  pending: 'pending',
  installing: 'installing',
  starting: 'starting',
  running: 'running',
  crashed: 'crashed',
  stopped: 'stopped',
  error: 'error',
} as const;

export interface HostedBotSummary {
  fileName: string;
  status: HostedBotSummaryStatus;
  /** @nullable */
  startCommand: string | null;
  restartCount: number;
  /** @nullable */
  errorMessage: string | null;
  /** @nullable */
  aiExplanation: string | null;
  /** @nullable */
  lastStartedAt: string | null;
}

export interface HostingSession {
  ticketId: number;
  ownerUsername: string;
  expiresAt: string;
  hostingDurationDays: number;
  hostedBot: HostedBotSummary | null;
}

export type HostOutcomeStatus = typeof HostOutcomeStatus[keyof typeof HostOutcomeStatus];


export const HostOutcomeStatus = {
  running: 'running',
  crashed: 'crashed',
  error: 'error',
} as const;

export interface HostOutcome {
  status: HostOutcomeStatus;
  message: string;
  /** @nullable */
  detail: string | null;
  /** @nullable */
  startCommand: string | null;
  /** @nullable */
  aiExplanation: string | null;
}